修改黄色

+++ CardLeap/CardLeapImplement/Contacter/SchoolTime/SendInfo2/PhotoView/PhotoBrowserViewController.m
@@ -221,8 +221,8 @@
 #pragma mark-----scrollViewDelegate
 -(void)scrollViewDidScroll:(UIScrollView *)scrollView
 {
-    currentIndex=(int)(scrollView.contentOffset.x/scrollView.frame.size.width);
-    indexLabel.text=[NSString stringWithFormat:@"%d / %d",currentIndex+1,count];
+    currentIndex=(NSInteger)(scrollView.contentOffset.x/scrollView.frame.size.width);
+    indexLabel.text=[NSString stringWithFormat:@"%ld / %d",currentIndex+1,count];
 }
 
 -(void)singleDoubleTap:(UITapGestureRecognizer*)recognizer
